requirements.txt
requests
pytest
pyyaml
jsonpath